api_auth/requirements.txt
wheel
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
PyJWT
